refactor: TopicCard 타입 분리 및 any 제거, 모든 참조 일관화
