+++ src/board/model/TopicCard.ts
@@ -1,0 +1,9 @@
+import { Timestamp } from "firebase/firestore"
+
+export type TopicCard = {
+  id: string
+  title: string
+  description: string
+  createdAt: Timestamp
+  createdBy: string
+}
